requirements.txt
line-bot-sdk
flask
gunicorn
redis
requests
numpy
